backend/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
psycopg2-binary
django-cors-headers

